Updated the folder structure for the main internal pages and their vms
