+++ app/src/main/java/com/fit2081/arrtish/id32896786/a1/internalpages/home/Home.kt
@@ -1,4 +1,4 @@
-package com.fit2081.arrtish.id32896786.a1.home
+package com.fit2081.arrtish.id32896786.a1.internalpages.home
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
@@ -11,7 +11,6 @@ import androidx.compose.runtime.*
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
