+++ app/src/main/java/com/fit2081/arrtish/id32896786/a1/internalpages/insights/InsightsViewModel.kt
@@ -1,5 +1,7 @@
-package com.fit2081.arrtish.id32896786.a1.home
+package com.fit2081.arrtish.id32896786.a1.internalpages.insights
 
+import android.content.Context
+import android.content.Intent
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -9,7 +11,7 @@ import com.fit2081.arrtish.id32896786.a1.databases.patientdb.PatientRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
-class HomeViewModel(private val repository: PatientRepository) : ViewModel() {
+class InsightsViewModel(private val repository: PatientRepository) : ViewModel() {
 
     private val _patient = MutableLiveData<Patient?>()
     val patient: LiveData<Patient?> = _patient
@@ -28,4 +30,28 @@ class HomeViewModel(private val repository: PatientRepository) : ViewModel() {
             }
         }
     }
+
+    fun sharingInsights(
+        context: Context,
+        userScores: Map<String, Float>,
+        totalScore: Float,
+        maxScore: Float
+    ) {
+        val shareText = buildString {
+            append("🌟 Insights: Food Score 🌟\n")
+            userScores.forEach { (category, score) ->
+                append("$category: %.2f/10\n".format(score))
+            }
+            append("\n🏆 Total Food Quality Score: %.2f/$maxScore".format(totalScore))
+        }
+
+        val intent = Intent(Intent.ACTION_SEND).apply {
+            type = "text/plain"
+            putExtra(Intent.EXTRA_TEXT, shareText)
+        }
+
+        val chooser = Intent.createChooser(intent, "Share your insights via:")
+        context.startActivity(chooser)
+    }
 }
+
